+++ FieldOpt/ERTWrapper/tests/test_eclsummaryreader.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include "ERTWrapper/eclsummaryreader.h"
 #include "ERTWrapper/ertwrapper_exceptions.h"
-#include "Utilities/tests/test_resource_example_file_paths.hpp"
+#include "Settings/tests/test_resource_example_file_paths.hpp"
 
 using namespace ERTWrapper::ECLSummary;
