+++ FieldOpt/Optimization/tests/constraints/test_reservoir_boundary.cpp
@@ -2,7 +2,7 @@
 #include "constraints/reservoir_boundary.h"
 #include "Optimization/tests/test_resource_cases.h"
 #include "Reservoir/tests/test_resource_grids.h"
-#include "Utilities/tests/test_resource_settings.hpp"
+#include "Settings/tests/test_resource_settings.hpp"
 
 
 namespace {
