+++ FieldOpt/Optimization/tests/objective/test_weightedsum.cpp
@@ -2,7 +2,7 @@
 #include <QString>
 #include "Optimization/objective/weightedsum.h"
 #include "Simulation/tests/test_resource_results.h"
-#include "Utilities/tests/test_resource_settings.hpp"
+#include "Settings/tests/test_resource_settings.hpp"
 
 
 using namespace Optimization::Objective;
