+++ FieldOpt/Settings/CMakeLists.txt
@@ -1,0 +1,37 @@
+cmake_minimum_required(VERSION 3.2)
+project(settings LANGUAGES CXX)
+
+add_library(settings
+        settings_exceptions.h
+        optimizer.cpp
+        simulator.cpp
+        model.cpp
+        settings.cpp
+        )
+add_library(fieldopt::settings ALIAS ${PROJECT_NAME})
+
+target_link_libraries(settings
+        PUBLIC Qt5::Core)
+
+target_include_directories(settings PUBLIC
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
+
+target_compile_features(settings
+        PUBLIC cxx_auto_type
+        PUBLIC cxx_range_for)
+
+# Unit tests
+find_package(GTest REQUIRED)
+include_directories(${GTEST_INCLUDE_DIRS})
+add_executable(test_settings
+        tests/test_settings.cpp
+        tests/test_settings_model.cpp
+        tests/test_settings_optimizer.cpp
+        tests/test_settings_simulator.cpp
+        )
+target_link_libraries(test_settings
+        fieldopt::settings
+        ${GTEST_BOTH_LIBRARIES}
+        ${CMAKE_THREAD_LIBS_INIT})
+
+add_test(NAME test_settings COMMAND $<TARGET_FILE:test_settings>)
