+++ FieldOpt/Settings/settings_exceptions.h
