+++ FieldOpt/Settings/tests/test_settings_model.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <QString>
 
-#include "Utilities/tests/test_resource_settings.hpp"
+#include "Settings/tests/test_resource_settings.hpp"
 
 using namespace Utilities::Settings;
