+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/ecldriverfilewriter.h
@@ -26,8 +26,8 @@
 #ifndef ECLDRIVERFILEWRITER_H
 #define ECLDRIVERFILEWRITER_H
 
-#include "Utilities/settings/settings.h"
-#include "Utilities/settings/simulator.h"
+#include "Settings/settings.h"
+#include "Settings/simulator.h"
 #include "Model/model.h"
 
 namespace Simulation { namespace SimulatorInterfaces {
