+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/flowdriverfilewriter.h
@@ -1,8 +1,8 @@
 #ifndef FIELDOPT_FLOWDRIVERFILEWRITER_H
 #define FIELDOPT_FLOWDRIVERFILEWRITER_H
 
-#include "Utilities/settings/settings.h"
-#include "Utilities/settings/simulator.h"
+#include "Settings/settings.h"
+#include "Settings/simulator.h"
 #include "Model/model.h"
 
 namespace Simulation {
