+++ FieldOpt/Simulation/simulator_interfaces/simulator.h
@@ -4,8 +4,8 @@
 #include <QString>
 #include "Model/model.h"
 #include "Simulation/results/results.h"
-#include "Utilities/settings/settings.h"
-#include "Utilities/settings/simulator.h"
+#include "Settings/settings.h"
+#include "Settings/simulator.h"
 #include "Simulation/execution_scripts/execution_scripts.h"
 
 namespace Simulation {
