+++ FieldOpt/Utilities/tests/test_execution.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include <tests/test_resource_example_file_paths.hpp>
+#include <Settings/tests/test_resource_example_file_paths.hpp>
 #include <Simulation/execution_scripts/execution_scripts.h>
 
 #include "execution.hpp"
